+++ resources/views/default/column/control_button.blade.php
@@ -6,6 +6,10 @@
             <i class="{{ $icon }}"></i>
         @endif
 
+        @if($image)
+            <img src="{{ $image }}">
+        @endif
+
         @if(!$hideText)
             {!! $text !!}
         @endif